pyproject.toml
[tool.poetry]
name = "worklogger"
version = "0.1.0"
description = ""
authors = ["Kirill Avdeev =-Жывтоне полосатЕ-= <morinehtar.qwerty@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
gitpython = "^3.1.43"
requests = "^2.32.3"
openai = "^1.39.0"
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
psycopg2 = "^2.9.9"
unidiff = "^0.7.5"
diff-match-patch = "^20230430"
pyqt5 = "^5.15.11"
qasync = "^0.27.1"
pytz = "^2024.1"
python-telegram-bot = "^21.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
